% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.package}
\alias{as.package}
\title{A simple alternative to \code{devtools::as.package}}
\usage{
as.package(x)
}
\arguments{
\item{x}{A package object to coerce}
}
\value{
A \code{package} object
}
\description{
Functionally identical to \code{devtools}' \code{as.package}, but without interactive
options for package creation.
}
\note{
Code inspired by \code{devtools} \code{load_pkg_description} with very minor
edits to further reduce \code{devtools} dependencies.
}
