% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_srcref.R
\name{as_list_of_srcref}
\alias{as_list_of_srcref}
\alias{as_list_of_srcref.environment}
\alias{as_list_of_srcref.list}
\title{Create an S3 list of srcref objects}
\usage{
as_list_of_srcref(x)

\method{as_list_of_srcref}{environment}(x)

\method{as_list_of_srcref}{list}(x)
}
\arguments{
\item{x}{A list or single srcref to coerce to a \code{list_of_srcref}}
}
\value{
A \code{list_of_srcref} class object
}
\description{
Create an S3 list of srcref objects
}
