% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{as_test_desc}
\alias{as_test_desc}
\alias{as_testthat_desc}
\title{Wrap object in test description derivation data}
\usage{
as_test_desc(x, type = "call")

as_testthat_desc(x)
}
\arguments{
\item{x}{A test description string to bind style data to}

\item{type}{A type class to attribute to the test description. Defaults to
\code{"call"}.}
}
\value{
A \code{test_description} subclass object with additional
\code{style} attribute indicating how the test description was derived.
}
\description{
Wrap object in test description derivation data

Adds "testthat" style
}
