% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_trace_df.R
\name{coverage_get_tests}
\alias{coverage_get_tests}
\title{Retrieve test traces from a coverage object}
\usage{
coverage_get_tests(coverage)
}
\arguments{
\item{coverage}{a \code{\link[covr]{covr}} coverage object}
}
\value{
A \code{list} of tests evaluated when using \code{covr}
}
\description{
Assumes the coverage object was produced while
\code{option(covr.record_tests = TRUE)}.
}
\seealso{
Other coverage_tests:
\code{\link[=coverage_check_has_recorded_tests]{coverage_check_has_recorded_tests()}},
\code{\link[=coverage_has_recorded_tests]{coverage_has_recorded_tests()}}
}
\concept{coverage_tests}
