% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.tools}
\alias{.tools}
\title{Loading unexported helpers from tools}
\format{
An object of class \code{list} of length 4.
}
\usage{
.tools
}
\description{
Loading unexported helpers from tools
}
\keyword{datasets}
