% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_srcref.R
\name{format.list_of_srcref}
\alias{format.list_of_srcref}
\title{Format a list_of_srcref object}
\usage{
\method{format}{list_of_srcref}(x, ..., full.names = FALSE, full.num = FALSE)
}
\arguments{
\item{x}{A \code{list_of_srcref} object}

\item{...}{Additional arguments unused}

\item{full.names}{A \code{logical} value indicating whether to use full file
paths when formatting \code{srcref}s.}

\item{full.num}{A \code{logical} value indicating whether to use all numeric
\code{srcref} components when formatting \code{srcref}s.}
}
\value{
A \code{character} vector of formatted strings
}
\description{
Format list_of_srcref as character
}
