% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_namespace_object_names}
\alias{get_namespace_object_names}
\title{Get namespace exports, filtering methods tables and definitions}
\usage{
get_namespace_object_names(ns)
}
\arguments{
\item{ns}{A namespace object}
}
\value{
The names of exported objects, filtering internal method tables and
metadata.
}
\description{
Get namespace exports, filtering methods tables and definitions
}
