% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_obj_namespace_name}
\alias{get_obj_namespace_name}
\title{Get namespace export namespace name}
\usage{
get_obj_namespace_name(ns, name)
}
\arguments{
\item{ns}{string or namespace object.}

\item{name}{string or name.}
}
\description{
For most objects, this will be identical to the namespace name provided, but
reexports will retain their originating package's namespace name. This helper
function helps to expose this name to determine which exports are reexports.
}
