% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSrcFilepath.R
\name{getSrcFilepath}
\alias{getSrcFilepath}
\title{Get the full path to the srcref file}
\usage{
getSrcFilepath(x)
}
\arguments{
\item{x}{A \code{srcref} or \code{list_of_srcref} object}
}
\value{
A \code{character} vector of source file paths.
}
\description{
Get the full path to the srcref file
}
