% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{is_srcref}
\alias{is_srcref}
\title{Test whether an object is a \code{srcref} object}
\usage{
is_srcref(x)
}
\arguments{
\item{x}{Any object}
}
\description{
Test whether an object is a \code{srcref} object
}
