% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcref_df.R
\name{match_containing_srcrefs}
\alias{match_containing_srcrefs}
\title{Match srcrefs against srcrefs that contain them}
\usage{
match_containing_srcrefs(l, r)
}
\arguments{
\item{l}{A \code{list_of_srcref} object}

\item{r}{A \code{list_of_srcref} object}
}
\value{
A \code{integer} vector of the first index in \code{r} that fully encapsulate
the respective element in \code{l}
}
\description{
Provided two lists of \code{srcref} objects, find the first \code{srcrefs} in \code{r} that
entirely encapsulate each respective \code{srcref} in \code{l}, returning a list of
indices of \code{srcref}s in \code{r} for each \code{srcref} in \code{l}.
}
