% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_namespace_name.R
\name{obj_namespace_name}
\alias{obj_namespace_name}
\title{Get namespace export namespace name}
\usage{
obj_namespace_name(x, ns)
}
\arguments{
\item{x}{A value to find within namespace \code{ns}}

\item{ns}{A package namespace}
}
\value{
A \code{character} string representing a namespace or similar
}
\description{
For most objects, this will be identical to the namespace name provided, but
reexports will retain their originating package's namespace name. This helper
function helps to expose this name to determine which exports are reexports.
}
