% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{package_check_has_keep_source}
\alias{package_check_has_keep_source}
\title{Verify that the package collection contains srcref information}
\usage{
package_check_has_keep_source(env)
}
\arguments{
\item{a}{package namespace environment or iterable collection of package
objects}
}
\description{
Test whether the package object collection contains srcref attributes.
}
