% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{srcref_expr}
\alias{srcref_expr}
\title{Parse the expression associated with a srcref}
\usage{
srcref_expr(ref)
}
\arguments{
\item{ref}{a \code{srcref}}
}
\value{
A parsed \code{srcref} object
}
\description{
Parse the expression associated with a srcref
}
