% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{srcref_namespace}
\alias{srcref_namespace}
\title{Get namespace of the \code{srcref} object}
\usage{
srcref_namespace(srcref)
}
\arguments{
\item{srcref}{object}
}
\value{
A \code{character} namespace name
}
\description{
Get namespace of the \code{srcref} object
}
