% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{srcref_str}
\alias{srcref_str}
\title{Parse the expression associated with a srcref}
\usage{
srcref_str(ref)
}
\arguments{
\item{ref}{a \code{srcref}}
}
\description{
Parse the expression associated with a srcref
}
