% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{test_description}
\alias{test_description}
\title{Parse a test description from the calling expression}
\usage{
test_description(x)
}
\arguments{
\item{x}{a unit test call stack or expression.}
}
\value{
A string that describes the test. If possible, this will be a written
description of the test, but will fall back to the test call as a string in
cases where no written description can be determined.
}
\description{
In the general case, a simple indicator of the source file and line number is
used as a test description. There are some special cases where more
descriptive information can be extracted:
}
\details{
\describe{
\item{\code{testthat}}{If the test used \code{\link[testthat]{test_that}},
then the description (\code{desc} parameter) is extracted and evaluated if
need be to produce a descriptive string. Nested calls to
\code{\link[testthat]{test_that}} currently return the outermost test
description, although this behavior is subject to change.
}
}
}
