% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{test_description_test_that}
\alias{test_description_test_that}
\title{Parse the test_that description from a test_that call}
\usage{
test_description_test_that(x, ...)
}
\arguments{
\item{x}{A test_that call object}

\item{...}{Additional arguments unused}
}
\description{
Parse the test_that description from a test_that call
}
