% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{test_srcrefs}
\alias{test_srcrefs}
\alias{test_srcrefs.coverage}
\title{Extract test srcref objects}
\usage{
test_srcrefs(x)

\method{test_srcrefs}{coverage}(x)
}
\arguments{
\item{x}{A \code{\link[covr]{package_coverage}} coverage object, from which
the test \code{srcref}s are extracted.}
}
\value{
A \code{list_of_srcref}
}
\description{
Extract test srcref objects
}
\examples{
options(covr.record_tests = TRUE)
pkg_path <- system.file("examplepkg", package = "covtracer")
cov <- covr::package_coverage(pkg_path)
test_srcrefs(cov)

}
\seealso{
as.data.frame.list_of_srcref

Other srcrefs:
\code{\link[=pkg_srcrefs]{pkg_srcrefs()}},
\code{\link[=trace_srcrefs]{trace_srcrefs()}}
}
\concept{srcrefs}
