% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{trace_srcrefs}
\alias{trace_srcrefs}
\alias{trace_srcrefs.coverage}
\title{Extract srcref objects from coverage object traces}
\usage{
trace_srcrefs(x)

\method{trace_srcrefs}{coverage}(x)
}
\arguments{
\item{x}{(\code{link[covr]{package_coverage}}) A \code{\link[covr]{covr}}
coverage object produced with \code{options(covr.record_tests = TRUE)}.}
}
\value{
A \code{list_of_srcref}
}
\description{
Extract srcref objects from coverage object traces
}
\seealso{
as.data.frame.list_of_srcref

Other srcrefs:
\code{\link[=pkg_srcrefs]{pkg_srcrefs()}},
\code{\link[=test_srcrefs]{test_srcrefs()}}
}
\concept{srcrefs}
