% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_trace_df.R
\name{coverage_has_recorded_tests}
\alias{coverage_has_recorded_tests}
\title{Test that the coverage object retains testing information}
\usage{
coverage_has_recorded_tests(coverage)
}
\arguments{
\item{coverage}{a \code{\link[covr]{covr}} coverage object}
}
\value{
A \code{logical} value, indicating whether the coverage object has
recorded tests, or \code{NA} when it does not appear to have traced any test
code.
}
\description{
Test whether the coverage object has expected fields produced when coverage
was captured with \code{option(covr.record_tests = TRUE)}.
}
\seealso{
Other coverage_tests: 
\code{\link{coverage_check_has_recorded_tests}()},
\code{\link{coverage_get_tests}()}
}
\concept{coverage_tests}
