% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{expr_str}
\alias{expr_str}
\title{Convert an expression, call or symbol to a single-line string}
\usage{
expr_str(ref)
}
\arguments{
\item{ref}{a \code{srcref}}
}
\value{
The given expression, formatted as a string with prefixes for
symbols and generics.
}
\description{
Convert an expression, call or symbol to a single-line string
}
