% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{flat_map_srcrefs}
\alias{flat_map_srcrefs}
\title{Map \code{srcrefs} over an iterable object, Filtering non-srcref results}
\usage{
flat_map_srcrefs(xs, ns = NULL)
}
\arguments{
\item{xs}{Any iterable object}
}
\description{
Map \code{srcrefs} over an iterable object, Filtering non-srcref results
}
