% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcrefs.R
\name{flat_map_srcrefs}
\alias{flat_map_srcrefs}
\title{Map \code{srcrefs} over an iterable object, Filtering non-srcref results}
\usage{
flat_map_srcrefs(xs, ns = NULL, breadcrumbs = character())
}
\arguments{
\item{xs}{Any iterable object}

\item{ns}{A \code{character} namespace name to attribute to objects in \code{xs}. If
\code{xs} objects themselves have namespaces attributed already to them, the
namespace will not be replaced.}

\item{breadcrumbs}{Recursive methods are expected to propegate a vector of
"breadcrumbs" (a character vector of namespace names encountered while
traversing the namespace used as a memory of what we've seen already),
which is used for short-circuiting recursive environment traversal.}
}
\value{
A \code{list} of \code{srcref}s
}
\description{
Map \code{srcrefs} over an iterable object, Filtering non-srcref results
}
