% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_empty_test_trace_tally}
\alias{new_empty_test_trace_tally}
\title{Build an empty covr-style test trace mapping}
\usage{
new_empty_test_trace_tally()
}
\value{
An empty test-trace matrix, as provided by \code{covr}
}
\description{
Build an empty covr-style test trace mapping
}
