% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{srcref_str}
\alias{srcref_str}
\title{Convert a srcref into a string}
\usage{
srcref_str(ref)
}
\arguments{
\item{ref}{a \code{srcref}}
}
\value{
A string representing the \code{srcref}
}
\description{
Convert a srcref into a string
}
