% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{test_description_test_that}
\alias{test_description_test_that}
\title{Parse the test description from a \code{test_that} call}
\usage{
test_description_test_that(x, ...)
}
\arguments{
\item{x}{A test_that call object}

\item{...}{Additional arguments unused}
}
\value{
A \code{character} description, parsed from a \code{test_that::test_that} call
}
\description{
Parse the test description from a \code{test_that} call
}
