% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_description.R
\name{test_description_test_that_describe}
\alias{test_description_test_that_describe}
\title{Parse the test description from a \code{describe} call}
\usage{
test_description_test_that_describe(x, ...)
}
\arguments{
\item{x}{A \code{test_that::describe} call object}

\item{...}{Additional arguments unused}
}
\value{
A \code{character} description, parsed from a \code{test_that::describe} call
}
\description{
Parse the test description from a \code{describe} call
}
