% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_srcref.R
\name{with_pseudo_srcref}
\alias{with_pseudo_srcref}
\title{For consistency, stub calls with srcref-like attributes}
\usage{
with_pseudo_srcref(call, file, lloc)
}
\arguments{
\item{call}{Any code object, most often a \code{call} object}

\item{file}{A filepath to bind as a \code{srcfile} object}

\item{lloc}{A \code{srcef}-like \code{lloc} numeric vector}
}
\value{
A \code{with_pseudo_srcref} object, mimicking the structure of \code{srcref}
}
\description{
Most relevant data can be traced to an existing srcref. However, some data,
such as test traces from coverage objects, are likely cleaned up and their
srcfiles deleted, causing a barrage of warnings any time these objects are
printed.
}
\details{
A \code{pseudo_srcref} adds in the \code{srcref} data but continues to preserve the
expression content. This allows these expression objects to be pretty-printed
like \code{srcref}s when included as a \code{list_of_srcref} \code{data.frame} column.
}
